# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Nunu Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "paths": {
...
        "/v1/knowledgeBase/createCategory": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识库模块"
                ],
                "summary": "新建知识库分类",
                "parameters": [
                    {
                        "description": "params",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CreateCategoryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/v1/knowledgeBase/createKB": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识库模块"
                ],
                "summary": "新建团队知识库",
                "parameters": [
                    {
                        "description": "params",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CreateKBRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.CreateKBResp"
                        }
                    }
                }
            }
        },
        "/v1/knowledgeBase/deleteCategory": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识库模块"
                ],
                "summary": "删除知识库分类",
                "parameters": [
                    {
                        "description": "params",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.DeleteCategoryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/v1/knowledgeBase/deleteKB": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识库模块"
                ],
                "summary": "删除知识库",
                "parameters": [
                    {
                        "description": "params",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.DeleteKBReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/v1/knowledgeBase/getCategoryListByKB": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识库模块"
                ],
                "summary": "获取知识库分类列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "知识库ID",
                        "name": "kb_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.CategoryData"
                        }
                    }
                }
            }
        },
        "/v1/knowledgeBase/getKBInfo": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识库模块"
                ],
                "summary": "获取知识库信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "知识库ID",
                        "name": "kb_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.GetKBInfoResp"
                        }
                    }
                }
            }
        },
        "/v1/knowledgeBase/getKBListByTeamId": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识库模块"
                ],
                "summary": "获取团队知识库列表",
                "parameters": [
                    {
                        "description": "params",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.GetKBListByTeamIdReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.GetKBListByTeamIdResp"
                        }
                    }
                }
            }
        },
        "/v1/knowledgeBase/getKBListByType": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识库模块"
                ],
                "summary": "获取知识库列表(私人知识库、公共知识库)",
                "parameters": [
                    {
                        "description": "params",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.GetKBListByTypeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.KBList"
                        }
                    }
                }
            }
        },
        "/v1/knowledgeBase/updateCategory": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识库模块"
                ],
                "summary": "更新知识库分类",
                "parameters": [
                    {
                        "description": "params",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.UpdateCategoryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/v1/knowledgeBase/updateKBName": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识库模块"
                ],
                "summary": "更新知识库名称",
                "parameters": [
                    {
                        "description": "params",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.UpdateKBNameReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "v1.CategoryData": {
            "type": "object",
            "properties": {
                "CategoryList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.CategoryView"
                    }
                }
            }
        },
...
        "v1.CreateCategoryReq": {
            "type": "object",
            "required": [
                "category_name",
                "kb_id"
            ],
            "properties": {
                "category_name": {
                    "type": "string"
                },
                "kb_id": {
                    "type": "integer"
                },
                "parent_id": {
                    "type": "integer"
                }
            }
        },
        "v1.CreateKBRequest": {
            "type": "object",
            "required": [
                "name",
                "team_id"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "team_id": {
                    "type": "integer"
                }
            }
        },
        "v1.CreateKBResp": {
            "type": "object",
            "properties": {
                "kb_id": {
                    "type": "integer"
                }
            }
        },
...
        "v1.DeleteCategoryReq": {
            "type": "object",
            "required": [
                "cid"
            ],
            "properties": {
                "cid": {
                    "type": "integer"
                }
            }
        },
        "v1.DeleteKBReq": {
            "type": "object",
            "required": [
                "kb_id"
            ],
            "properties": {
                "kb_id": {
                    "type": "integer"
                }
            }
        },
...
        "v1.GetKBInfoResp": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "kb_id": {
                    "type": "integer"
                },
                "kb_name": {
                    "type": "string"
                },
                "kb_type": {
                    "type": "string"
                },
                "team_id": {
                    "type": "integer"
                },
                "team_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "v1.GetKBListByTeamIdReq": {
            "type": "object",
            "required": [
                "team_id"
            ],
            "properties": {
                "page_index": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页大小",
                    "type": "integer"
                },
                "team_id": {
                    "type": "integer"
                }
            }
        },
        "v1.GetKBListByTeamIdResp": {
            "type": "object",
            "properties": {
                "kb_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.GetKBInfoResp"
                    }
                },
                "page_index": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页大小",
                    "type": "integer"
                },
                "total_count": {
                    "description": "总记录数",
                    "type": "integer"
                }
            }
        },
        "v1.GetKBListByTypeReq": {
            "type": "object",
            "properties": {
                "kb_type": {
                    "type": "string"
                }
            }
        },
...
        "v1.KBList": {
            "type": "object",
            "properties": {
                "kb_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.KbKnowledgeBaseView"
                    }
                }
            }
        },
...
        "v1.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "v1.UpdateCategoryReq": {
            "type": "object",
            "required": [
                "category_name",
                "cid"
            ],
            "properties": {
                "category_name": {
                    "type": "string"
                },
                "cid": {
                    "type": "integer"
                }
            }
        },
        "v1.UpdateKBNameReq": {
            "type": "object",
            "required": [
                "kb_id",
                "name"
            ],
            "properties": {
                "kb_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "vo.CategoryView": {
            "type": "object",
            "properties": {
                "category_name": {
                    "description": "分类名",
                    "type": "string"
                },
                "children": {
                    "description": "子分类",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.CategoryView"
                    }
                },
                "cid": {
                    "description": "分类ID",
                    "type": "integer"
                },
                "kb_id": {
                    "description": "知识库ID",
                    "type": "integer"
                },
                "level": {
                    "description": "层级",
                    "type": "integer"
                },
                "parent_id": {
                    "description": "父分类ID",
                    "type": "integer"
                }
            }
        },
        "vo.KbKnowledgeBaseView": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "kb_id": {
                    "type": "integer"
                },
                "kb_name": {
                    "type": "string"
                },
                "kb_type": {
                    "description": "知识库类型：私人/公共/团队",
                    "type": "string"
                },
                "team_id": {
                    "description": "团队ID，可能为空",
                    "type": "integer"
                },
                "team_name": {
                    "description": "团队名称",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "description": "用户ID，可能为空（私人知识库）",
                    "type": "string"
                },
                "user_name": {
                    "description": "用户昵称",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}